Refactor: move user dtos to separate package and improve semantics
