--- src/main/java/com/chat/chat_api/user/CreateUser.java
@@ -1,5 +1,0 @@
-package com.chat.chat_api.user;
-
-public record CreateUser(String username) {
-    
-}
