+++ src/main/java/com/chat/chat_api/user/dto/CreateUserRequestDTO.java
@@ -1,0 +1,5 @@
+package com.chat.chat_api.user.dto;
+
+public record CreateUserRequestDTO(String username) {
+    
+}
