+++ src/main/java/com/chat/chat_api/user/dto/UserDTO.java
@@ -1,10 +1,11 @@
-package com.chat.chat_api.user;
+package com.chat.chat_api.user.dto;
 
 import java.util.List;
 
 import com.chat.chat_api.chatroom.Chatroom;
+import com.chat.chat_api.user.User;
 
-public record UserDTO(Long id, String username, List<Long> chatIds) {
+public record UserDTO(Long userId, String username, List<Long> chatIds) {
 
     public static UserDTO toDto(User user){
         List<Long> ids = user.getChats()
